# api/lib/routes/opus-requirements-id-subscriptors-userid-delete.ts
import { Router, Request, Response, NextFunction } from 'express';
import { runCommand } from '../utils/bus/send-command';
import validateRequirement from '../utils/middlewares/validate-requirement';
import { RequirementSubscriptor } from '@jiku/models';
import logger from '../logger';

const router: Router = Router();

function validateSelfUnsubscribe(req: Request, res: Response, next: NextFunction) {
  if (req.params.userId !== req.user.id) {
    return res.status(403).json({ code: 'access_denied', message: 'Access denied' });
  }
  return next();
}

function validateSubscriptionExists(req: Request, res: Response, next: NextFunction) {
  return RequirementSubscriptor.findOne({
    where: { requirementId: req.requirement.id, userId: req.params.userId },
  })
    .then((subscription) => {
      if (!subscription) {
        return res.status(404).json({ code: 'subscription_not_found', message: 'Subscription not found' });
      }
      req.subscription = subscription;
      return next();
    })
    .catch((error: Error) => {
      logger.error(`DELETE /opus/requirements/:reqid/subscriptors/:userId validateSubscriptionExists error: ${error.message}`);
      return res.status(500).json({ code: 'internal_error', message: 'Internal server error' });
    });
}

/** El borrado lo hace core; que solo pueda desuscribirse a sí mismo lo valida la api. */
async function deleteSubscription(req: Request, res: Response) {
  const ok = await runCommand(
    res,
    `requirements.${req.requirement.id}.subscriptors.${req.params.userId}.delete`,
    {}
  );
  if (!ok) {
    return;
  }

  return res.status(200).json({});
}

/**
 * Esta ruta NO lleva `validateRequirementIsPublic`, a diferencia del resto de la superficie, y
 * es deliberado: solo borra la suscripcion DEL PROPIO caller (`validateSelfUnsubscribe`), asi
 * que no revela nada del requisito. Bloquearla dejaria atrapado a quien se suscribio cuando el
 * requisito era publico y despues paso a interno: seguiria recibiendo la suscripcion sin poder
 * darla de baja.
 */
router.delete('/opus/requirements/:reqid/subscriptors/:userId',
  validateSelfUnsubscribe,
  validateRequirement,
  validateSubscriptionExists,
  deleteSubscription
);

export default router;