# docs/apis/api.yaml
openapi: 3.0.3
...
info:
  title: Jiku api — contrato HTTP
  version: 1.0.0
  description: |
    Única puerta HTTP del producto. Autentica contra Zitadel, autoriza por rol y por entidad,
    **lee la base de datos directamente** y convierte toda mutación en un comando publicado en
    NATS que atiende `core`.

    **Las escrituras pasan por el bus.** Los endpoints marcados con `x-bus-command` publican un
    comando y esperan respuesta. Eso agrega **tres** modos de falla a todos ellos:

    - **`503` (`service_unavailable`) — problema de despliegue.** No hay **ningún suscriptor** del
      subject. El server contesta *"no responders"* **en milisegundos**, no al vencer el timeout.
      La operación **no ocurrió**, así que reintentar es seguro.
    - **`504` (`gateway_timeout`) — problema de performance.** Había alguien escuchando, pero la
      respuesta no llegó dentro de `NATS_REQUEST_TIMEOUT_MS` (default 5000 ms). **La operación
      pudo haber ocurrido**: sin JetStream no hay acuse, así que reintentar a ciegas puede
      duplicar.
    - **Los códigos de error que devuelve core**, traducidos a HTTP por `httpStatusFor`.

    Los dos primeros los genera la api en su `catch`; **`gateway_timeout` NO es un `errorCode` de
    core** y por eso **no está en `STATUS_BY_ERROR_CODE`**. Separarlos es lo que permite distinguir
    "no desplegué core" de "core está lento" sin mirar los logs. El contrato del bus está en
    [core.yaml](./core.yaml).

...

...

...

...
    >
...

...

...
    produce hoy.
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /opus/requirements/{reqid}/subscriptors/{userId}:
    delete:
      tags: [opus]
      summary: Desuscribe un usuario de un requisito
      description: |
        Publica `requirements.{id}.subscriptors.{userId}.delete`. **Un `external-user` solo puede
        desuscribirse a sí mismo.**

        **Es el único endpoint de `/opus/requirements/{reqid}` que NO recorta por visibilidad**,
        y es deliberado: solo borra la suscripción del propio caller, así que no revela nada del
        requisito, y bloquearlo dejaría atrapado a quien se suscribió cuando el requisito era
        público y después pasó a interno.
      x-roles: [external-user]
      x-bus-command: requirements.{id}.subscriptors.{userId}.delete
      parameters:
        - $ref: '#/components/parameters/ReqIdPath'
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Acknowledgement' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Intento de desuscribir a otro usuario, o sin permiso sobre el proyecto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: El requisito o la suscripción no existen
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: subscription_not_found, message: Subscription not found }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/BusUnavailable' }
        '504': { $ref: '#/components/responses/GatewayTimeout' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT RS256 emitido por Zitadel, verificado contra su JWKS
        (`{IDENTITY_URL}/oauth/v2/keys`), resincronizado con reintentos cuando aparece un `kid`
        desconocido.

        **Solo por header.** El fallback por query param (`?jwt=`) se eliminó: los query params
        quedan en logs de acceso, historial y `Referer`.

        Los roles salen del claim `urn:zitadel:iam:org:project:roles`, como las claves del objeto.
        Roles del producto: `admin`, `user`, `external-user`.

        > **REQ-007 (S-034) — `- ELIMINADO`.** Acá decía: *"El `sub` tiene que existir en la
        > tabla `users` o la respuesta es 401 `user_not_found`"*. **Ese 401 ya no existe en
        > ninguna de las 61 rutas.**
        >
        > `validateToken` **dejó de hacer `User.findByPk(sub)`** y arma `req.user` con
        > `{ id, name, username, email, roles }` **del claim ya verificado** — que es la fuente
        > que ADR-007 declara autoritativa (*"los roles DEBEN leerse del claim […] NO SE DEBEN
        > almacenar roles en la base ni derivarlos de otra fuente"*). Con eso:
        >
        > - una identidad autenticada **opera aunque no tenga fila previa**, y la fila queda
        >   creada la primera vez que ejecuta un comando (`core` la espeja desde el sobre)
        > - **se elimina un `SELECT` por request** en toda la superficie HTTP
        > - `POST /api/auth/present` **deja de necesitar su excepción**: ya no hay chequeo del
        >   que eximirlo
        >
        > **El `401 unauthorized` (sin token / token inválido) NO cambia.**
  parameters:
...
    ReqIdPath:
      name: reqid
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    Unauthorized:
      # REQ-007 (S-034): `- ELIMINADO` el ejemplo `userNotFound`. El `sub` ya NO tiene que estar
      # en `users`: `req.user` se arma del claim. Queda una sola causa de 401.
      description: Sin token o token inválido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            unauthorized:
              value: { code: unauthorized, message: Unauthorized }
...
    InternalError:
      description: Error inesperado. El detalle va al log, no a la respuesta
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: internal_error, message: Internal error }
    BusUnavailable:
      description: |
        **No hay ningún suscriptor del subject:** el server contesta *"no responders"* **en
        milisegundos**, no al vencer el timeout. Es la señal de un problema de **despliegue**.

        **La operación no ocurrió**, y sin JetStream no hay reintento — así que **reintentar es
        seguro**.

        No cubre el timeout: eso es `504` (`GatewayTimeout`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            code: service_unavailable
            message: El servicio no está disponible en este momento
    GatewayTimeout:
      description: |
        **La respuesta no llegó dentro de `NATS_REQUEST_TIMEOUT_MS`** (default 5000 ms). Había
        alguien escuchando y no contestó a tiempo: es la señal de un problema de **performance**,
        no de despliegue.

        **LA OPERACIÓN PUDO HABER OCURRIDO.** Es el modo de falla más confuso del sistema: sin
        JetStream no hay acuse ni idempotencia (ADR-002), así que core puede haber escrito y la
        respuesta haberse perdido. **Reintentar a ciegas puede duplicar.** El `message` no invita
        a reintentar, a propósito.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            code: gateway_timeout
            message: La operación tardó demasiado
  schemas:
...
    Error:
      type: object
      description: |
        La forma de error de toda la api. `message` es **texto de interfaz**: los frontends lo
        muestran tal cual al usuario. Están mezclados entre inglés y español.
      required: [code, message]
      properties:
        code:
          type: string
          description: Código estable. Los frontends hacen `switch` sobre él
        message: { type: string }
        remainingMinutes:
          type: integer
          description: |
            Solo en `daily_limit_exceeded`. Se recupera parseando el mensaje con un regex, porque
            el formato de respuesta del bus no tiene dónde poner datos extra de un error. Es una
            solución de transición
    Acknowledgement:
      type: object
      description: Acuse de una edición que no devuelve el recurso
      properties:
        code: { type: string }
        message: { type: string }

    # ---- enums ----
...